+++ components/todoItem.js
@@ -1,10 +1,14 @@
 import React from 'react';
-import { StyleSheet, Text, TouchableOpacity } from 'react-native';
+import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import { MaterialIcons } from '@expo/vector-icons';
 
 export default function TodoItem({ item, pressHandler }) {
   return (
     <TouchableOpacity onPress={ () => pressHandler(item._id)}>
-      <Text style={styles.item}>{item.text}</Text>
+      <View style={styles.item}>
+        <MaterialIcons name="delete" size={18} color="darkred" />
+        <Text style={styles.itemText}>{item.text}</Text>
+      </View>
     </TouchableOpacity>
   );
 }
@@ -16,5 +20,9 @@ const styles = StyleSheet.create({
     borderColor: 'lightgray',
     borderWidth: 1,
     borderRadius: 10,
+    flexDirection: "row"
+  },
+  itemText: {
+    marginLeft: 10
   }
 });
