+++ screens/home.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+export default function Home() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.titleText}>Home Screen</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    padding: 24
+  },
+  titleText: {
+    fontFamily: 'nunito-bold',
+    fontSize: 18
+  }
+});
