+++ screens/shopingList.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+export default function ShopiingList() {
+  return (
+    <View style={styles.container}>
+      <Text>ShopiingList Screen</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+    container: {
+        padding: 24
+    }
+});
